Muutin levelien käsittelyä niin että leveli voi olla desimaaliluku, eli vaikka -4.5 [m], kun ennen se voi olla vain positiivinen kokonaisluku. Muutos näkyy NFmiDrawParamFactory::CreateFileName -metodissa.

+++ source/NFmiDrawParamFactory.cpp
@@ -180,7 +180,7 @@ std::string NFmiDrawParamFactory::CreateFileName(NFmiDrawParam* drawParam, bool 
 					int levelTypeId = level.LevelTypeId();
 					fileName += NFmiStringTools::Convert(levelTypeId);
 					fileName += "_";
-					int levelValue = level.LevelValue();
+					float levelValue = level.LevelValue();
 					fileName += NFmiStringTools::Convert(levelValue);
 				}
 			}
